Refactor "Modal" into it's own file
